At Troels' proposal I've removed some checks for naturalnum.
Only checkset actually has to make sure that we get a proper set, and a
lot of the other functions check it implicit.
By doing it this way we can now successfully run

  ismember(X, [z, s(z)], yes)

without entering an infinite loop.

If we choose to do this for our assignment we should probably update our
report to reflect our assumptions about the input to ismember and less.